all: implement API and interface for attack functionality
When the user click "Attack" it will call the API to run the load testing.
If there is load testing currently, it will return with an error.

On success, the result of load testing will be stored on directory
defined in Environment.ResultsDir with file named

	<HttpTarget.ID>.<date_time>.<RPS>x<Duration>s.<ResultsSuffix>.bin

This file contains the vegeta.Results.