all: load pass attack results and implement function to get attack result
When the service started, it will load all previous attack results
from directory Environment.ResultsDir.
It will only scan the file name and append it to HttpTarget.Results
due to the size and time to load one of them can take time.

Through the web interface, user can click "Show" button to load
the result and display it on the screen.