all: allow client to customize request method, path, and type
Previously, the HttpTarget Method, Path, and RequestType is fixed.
Once its declared, the Run method will use the defined values to
generate and call HTTP request.

This changes add IsCustomizable field on HttpTarget. If its true,
client or web user interface can modify the request method, path, and
type.  The values send by client to server after changes will be used
instead of the fixed values.

This changes also make the Run handler on HttpTarget to be optional.
If its not defined, it will generated the HTTP request from the method,
path, and type passed to server.