www: refactoring using TypeScript and wui library
Previously, the web user interface is written in pure, single JavaScript
file.  The LOC is short but its become hard to maintenance, especially
when there is a change in HTML layout or on the response format.

This changes rewrite the interface to use TypeScript in order to easily
maintenance.  The generated JavaScript is loaded using module [1].

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Modules