all: add option to set description (Hint) on any inputs element
This changes refactor the KeyValue from map[string]string into
map[string]FormInput, where FormInput is a struct with Label, Hint,
Kind, Value, Max, and Min.

With the available of Hint field, the web user interface can render it
as description of input.

This changes also make all input hint to be displayed on the first render.