all: merge all workers for related to development to internal/cmd
Previously, we have a workers to recompile TypeScript and regenerate
HTML files from adoc inside the Trunks type.  The workers will run
if the environment variable TRUNKS_DEV set to non-zero.

This changes move those workers to the internal/cmd/trunks-example,
because those workers are related for development only not for running
Trunks server.