Release gorankusu v0.5.0
This release rename the project from "trunks" to "gorankusu".

The original idea of "trunks" is because the core library that we use
for load testing is named vegeta (from Dragon Ball), and Vegeta has a
son named Trunks.
In English, trunks also have multiple meanings.

In order to have a unique name, we rename the project to "gorankusu",
which is a combination of "go" (the main programming language that
built the application) and "torankusu" the Hepburn of "Trunks".

=== New features

* all: allow submit free form request body in HTTPTarget
* all: implement form input file
* all: add type to customize how to dump HTTP request and response
* all: support parameter binding in HTTP Path

=== Enhancements

* _www: check HTTP response status greater or equal 400